module7-task1: - рефакторинг редьюсера
